Better organization of objects and resolving a circular import
